+++ packages/adapter-fela/tests/FelaAesthetic.test.ts
@@ -28,6 +28,7 @@ import {
   SYNTAX_MEDIA_QUERY_NESTED,
   SYNTAX_KEYFRAMES_INLINE,
   SYNTAX_FONT_FACES_INLINE,
+  SYNTAX_RAW_CSS,
 } from '../../../tests/mocks';
 import { cleanStyles } from '../../../tests/helpers';
 
@@ -36,13 +37,13 @@ describe('FelaAesthetic', () => {
 
   function renderAndTest(
     styles: any,
-    expStyles: any,
-    expClassName: string,
+    expStyles: any = {},
+    expClassName: string = '',
     global: boolean = false,
   ) {
     const nativeSheet = global
       ? instance.syntax.convertGlobalSheet(styles).toObject()
-      : instance.syntax.convertStyleSheet(styles).toObject();
+      : instance.syntax.convertStyleSheet(styles, 'fela').toObject();
     // @ts-ignore Allow access
     const styleSheet = instance.processStyleSheet(nativeSheet, 'fela');
 
@@ -301,6 +302,10 @@ describe('FelaAesthetic', () => {
         },
         'a b c',
       );
+    });
+
+    it('handles raw CSS', () => {
+      renderAndTest(SYNTAX_RAW_CSS);
     });
   });
 });