+++ packages/core/src/Aesthetic.tsx
@@ -7,6 +7,7 @@ import deepMerge from 'extend';
 import { Omit } from 'utility-types';
 import isObject from './helpers/isObject';
 import stripClassPrefix from './helpers/stripClassPrefix';
+import StyleSheetManager from './StyleSheetManager';
 import UnifiedSyntax from './UnifiedSyntax';
 import {
   ClassName,
@@ -56,6 +57,8 @@ export default abstract class Aesthetic<
   themes: { [themeName: string]: Theme } = {};
 
   protected appliedGlobals: boolean = false;
+
+  protected sheetManager: StyleSheetManager | null = null;
 
   constructor(options: Partial<AestheticOptions> = {}) {
     this.options = {
@@ -115,14 +118,15 @@ export default abstract class Aesthetic<
 
     this.applyGlobalStyles();
 
-    const styleSheet = this.processStyleSheet(
-      this.syntax.convertStyleSheet(this.getStyleSheet(styleName)).toObject(),
-      styleName,
-    );
+    const baseSheet = this.syntax.convertStyleSheet(this.getStyleSheet(styleName), styleName);
+    const styleSheet = this.processStyleSheet(baseSheet.toObject(), styleName);
 
-    this.cache[styleName] = styleSheet;
+    this.cache[styleName] = {
+      ...styleSheet,
+      ...baseSheet.classNames,
+    } as SheetMap<ParsedBlock>;
 
-    return styleSheet;
+    return this.cache[styleName];
   }
 
   /**
@@ -381,6 +385,19 @@ export default abstract class Aesthetic<
     }
 
     return styleSheet;
+  }
+
+  /**
+   * Return a native style sheet manager used for injecting CSS.
+   */
+  protected getStyleSheetManager(): StyleSheetManager {
+    if (this.sheetManager) {
+      return this.sheetManager;
+    }
+
+    this.sheetManager = new StyleSheetManager();
+
+    return this.sheetManager;
   }
 
   /**