+++ packages/adapter-aphrodite/CHANGELOG.md
@@ -1,3 +1,17 @@
+# 3.3.0
+
+#### 🚀 New
+
+- Added support for raw CSS declarations in a style sheet.
+
+#### 🐞 Fixed
+
+- Added missing `@babel/runtime` package.
+
+#### 🛠 Internal
+
+- Updated `aphrodite` peer requirement to 2.3.
+
 # 3.2.0 - 2019-02-14
 
 #### 🛠 Internal