+++ packages/adapter-typestyle/CHANGELOG.md
@@ -1,3 +1,13 @@
+# 2.3.0
+
+#### 🚀 New
+
+- Added support for raw CSS declarations in a style sheet.
+
+#### 🐞 Fixed
+
+- Added missing `@babel/runtime` package.
+
 # 2.2.0 - 2019-02-14
 
 #### 🛠 Internal